requirements.txt
asgiref==3.6.0
dj-database-url==1.0.0
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.1
gunicorn==20.1.0
psycopg2==2.9.3
pymongo==3.12.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.2.0
